LINE/ChatBot/LINE/requirements.txt
line-bot-sdk==1.14.0
Flask==1.0.2
requests==2.21.0